+++ pkg/auth/BUILD
