+++ pkg/auth/authorizer/doc.go
@@ -15,4 +15,4 @@ limitations under the License.
 */
 
 // Package authorizer contains implementations for pkg/auth/authorizer interfaces
-package authorizer // import "k8s.io/kubernetes/plugin/pkg/auth/authorizer"
+package authorizer // import "k8s.io/kubernetes/pkg/auth/authorizer"
