+++ pkg/auth/authorizer/rbac/bootstrappolicy/namespace_policy.go
