+++ pkg/auth/authorizer/rbac/bootstrappolicy/testdata/cluster-role-bindings.yaml
