+++ pkg/auth/authorizer/rbac/bootstrappolicy/testdata/namespace-role-bindings.yaml
