+++ pkg/auth/authorizer/rbac/bootstrappolicy/testdata/namespace-roles.yaml
