+++ pkg/auth/doc.go
